requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.14
filelock==3.12.1
frozenlist==1.3.3
fsspec==2023.6.0
humanize==4.6.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lightning-utilities==0.8.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
Pygments==2.15.1
pytorch-lightning==2.0.3
PyYAML==6.0
requests==2.31.0
rich==13.4.1
scikit-learn==1.2.2
scipy==1.10.1
shellingham==1.5.0.post1
SimpleITK==2.2.1
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torchio==0.18.91
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
wrapt==1.15.0
yarl==1.9.2
